Key infrastructure added for attributes and affixes
It took a long time to find a suitable design for the affix system,
which required a flexible re-design for the attribute system as well.
It's far from complete (currently only weapons have affixes, only for
simple attributes, and currently the attributes are not actually being
used to affect gameplay) but a lot of the tricky design questions have
been answered with this commit. Key changes involve the following
classes: AttributeAggregator, UDictionary, EnumDictionary, Weapon,
AffixDatabase, and Inventory.